test(llm): split tools playwright suite
